Move TestMwsAccWorkspaceID_UnknownFromMwsWorkspaces alongside other mws_workspaces tests
Folds the new account-level integration test into mws/mws_workspaces_test.go
next to TestMwsAccWorkspaces, which has the same fixture shape. Removes the
standalone file in internal/acceptance/.
